# openapi.yml
openapi: 3.0.3
info:
  title: SeControle - Controle Financeiro Familiar API
  description: API para gerenciar transações, receitas, metas e cadastros de apoio do controle financeiro familiar.
  version: "3.0.0"
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
paths:
...
  /pessoas:
    get:
      tags: [Pessoas]
      summary: Lista pessoas que compartilham alguma casa com o usuário autenticado
      description: Idêntico a /pessoas/relacionadas — mantido por retrocompatibilidade.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AtivoFiltro"
      responses:
        "200":
          description: Lista de pessoas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pessoa"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
    post:
      tags: [Pessoas]
      summary: Cria uma pessoa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PessoaInput"
      responses:
        "201":
          description: Pessoa criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
...
  /pessoas/{id}:
    get:
      tags: [Pessoas]
      summary: Busca uma pessoa pelo id
      description: Retorna a pessoa se for a própria pessoa autenticada ou compartilhar alguma casa com ela.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Pessoa encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    put:
      tags: [Pessoas]
      summary: Atualiza uma pessoa
      description: Requer que o usuário autenticado esteja editando os próprios dados.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PessoaInput"
      responses:
        "200":
          description: Pessoa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pessoa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # -------------------------------------------------------------------
  # Casas
  # -------------------------------------------------------------------
  /casas:
    get:
      tags: [Casas]
      summary: Lista casas
      description: Todas as casas do sistema são visíveis a qualquer usuário autenticado.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AtivoFiltro"
      responses:
        "200":
          description: Lista de casas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Casa"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
    post:
      tags: [Casas]
      summary: Cria uma casa
      description: >
        Qualquer usuário autenticado pode criar uma casa. O criador é
        automaticamente adicionado como admin da casa criada.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NomeInput"
      responses:
        "201":
          description: Casa criada (o usuário autenticado já é admin dela)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Casa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "409":
          description: Já existe uma casa com esse nome
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
  /casas/{id}:
    put:
      tags: [Casas]
      summary: Atualiza uma casa
      description: Requer que o usuário autenticado seja admin da casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NomeInput"
      responses:
        "200":
          description: Casa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Casa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
...
  /casas/{casaId}/pessoas:
    get:
      tags: [Casa Pessoas]
      summary: Lista membros de uma casa
      description: Requer que o usuário autenticado participe desta casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CasaId"
      responses:
        "200":
          description: Lista de membros com dados da pessoa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CasaPessoa"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
    post:
      tags: [Casa Pessoas]
      summary: Associa uma pessoa a uma casa
      description: Requer que o usuário autenticado seja admin desta casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CasaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CasaPessoaInput"
      responses:
        "201":
          description: Pessoa associada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasaPessoa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "409":
          description: Pessoa já associada a esta casa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
  /casas/{casaId}/pessoas/{pessoaId}/papel:
    patch:
      tags: [Casa Pessoas]
      summary: Atualiza o papel de uma pessoa na casa
      description: Requer que o usuário autenticado seja admin desta casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CasaId"
        - $ref: "#/components/parameters/PessoaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [papel]
              properties:
                papel:
                  type: string
                  enum: [admin, membro]
                  example: admin
      responses:
        "200":
          description: Papel atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CasaPessoa"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /casas/{casaId}/pessoas/{pessoaId}:
    delete:
      tags: [Casa Pessoas]
      summary: Remove a associação de uma pessoa a uma casa
      description: Requer que o usuário autenticado seja admin desta casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CasaId"
        - $ref: "#/components/parameters/PessoaId"
      responses:
        "204":
          description: Associação removida
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # -------------------------------------------------------------------
  # Percentual de Custeio
  # -------------------------------------------------------------------
...
  /cartoes-contas/{cartaoId}/visibilidade:
    get:
      tags: [Cartões e Contas]
      summary: Lista o compartilhamento de um cartão/conta com casas
      description: >
        O titular vê todas as entradas de visibilidade do cartão. Um membro de
        casa que não é o titular só vê a entrada referente à própria casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CartaoId"
      responses:
        "200":
          description: Lista de entradas de visibilidade
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartaoCasaVisibilidade"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    post:
      tags: [Cartões e Contas]
      summary: Define o compartilhamento de um cartão/conta com uma casa
      description: Requer que o usuário autenticado seja o titular do cartão.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CartaoId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartaoCasaVisibilidadeInput"
      responses:
        "201":
          description: Entrada de visibilidade criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartaoCasaVisibilidade"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          description: Já existe uma entrada de visibilidade para esta casa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Erro"
  /cartoes-contas/{cartaoId}/visibilidade/{casaId}:
    patch:
      tags: [Cartões e Contas]
      summary: Alterna o compartilhamento de um cartão/conta com uma casa
      description: Requer que o usuário autenticado seja o titular do cartão.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CartaoId"
        - $ref: "#/components/parameters/CasaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [compartilhado]
              properties:
                compartilhado:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Entrada de visibilidade atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartaoCasaVisibilidade"
        "400":
          $ref: "#/components/responses/ErroValidacao"
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Cartões e Contas]
      summary: Remove o compartilhamento de um cartão/conta com uma casa
      description: >
        Requer que o usuário autenticado seja o titular do cartão. Remover a
        entrada volta o cartão ao padrão privado para aquela casa.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CartaoId"
        - $ref: "#/components/parameters/CasaId"
      responses:
        "204":
          description: Entrada de visibilidade removida
        "401":
          $ref: "#/components/responses/NaoAutenticado"
        "403":
          $ref: "#/components/responses/SemPermissao"
        "404":
          $ref: "#/components/responses/NaoEncontrado"

  # -------------------------------------------------------------------
  # Compras
  # -------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido pelo Supabase Auth após login
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
    CasaId:
      name: casaId
      in: path
      required: true
      schema:
        type: integer
      example: 1
    PessoaId:
      name: pessoaId
      in: path
      required: true
      schema:
        type: integer
      example: 7
...
    CartaoId:
      name: cartaoId
      in: path
      required: true
      schema:
        type: integer
      example: 2
...
    AtivoFiltro:
      name: ativo
      in: query
      required: false
      description: Filtra pelo status ativo/inativo
      schema:
        type: boolean
...
  responses:
    ErroValidacao:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
    NaoEncontrado:
      description: Registro não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
    NaoAutenticado:
      description: Token não fornecido ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
...
    SemPermissao:
      description: Usuário autenticado, mas sem permissão para esta ação
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Erro"
  schemas:
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: Registro não encontrado
    NomeInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          example: Energia
    PessoaInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          example: Erica
        email:
          type: string
          format: email
          nullable: true
          example: erica@email.com
    Pessoa:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Erica
        email:
          type: string
          format: email
          nullable: true
          example: erica@email.com
        supabase_user_id:
          type: string
          format: uuid
          nullable: true
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        ativo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    Casa:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Minha Casa
        ativo:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
    CasaPessoaInput:
      type: object
      required: [pessoa_id]
      properties:
        pessoa_id:
          type: integer
          example: 7
        papel:
          type: string
          enum: [admin, membro]
          default: membro
          example: admin
    CasaPessoa:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pessoa_id:
          type: integer
          example: 7
        papel:
          type: string
          enum: [admin, membro]
          example: admin
        created_at:
          type: string
          format: date-time
        nome:
          type: string
          example: Erica
        email:
          type: string
          format: email
          nullable: true
          example: erica@email.com
        ativo:
          type: boolean
          example: true
...
    CartaoCasaVisibilidadeInput:
      type: object
      required: [casa_id]
      properties:
        casa_id:
          type: integer
          example: 1
        compartilhado:
          type: boolean
          default: false
          example: true
    CartaoCasaVisibilidade:
      allOf:
        - $ref: "#/components/schemas/CartaoCasaVisibilidadeInput"
        - type: object
          properties:
            id:
              type: integer
              example: 1
            cartao_id:
              type: integer
              readOnly: true
              example: 2
            created_at:
              type: string
              format: date-time
...